+++ src/__screenshot_tests__/feedback-screenshot-test.tsx
@@ -4,7 +4,11 @@ import {MOVISTAR_SKIN, VIVO_SKIN, O2_SKIN} from '../skins/constants';
 
 const testableSkins = [MOVISTAR_SKIN, VIVO_SKIN, O2_SKIN];
 const testableDevices: Array<Device> = ['MOBILE_IOS', 'DESKTOP'];
-const feedbackTypes = ['success', 'error', 'info'];
+const feedbackTypes = [
+    'successfeedbackscreen--success',
+    'errorfeedbackscreen--error',
+    'infofeedbackscreen--info',
+];
 
 const cases: Array<[string, string, Device]> = [];
 for (const skin of testableSkins) {
@@ -17,7 +21,7 @@ for (const skin of testableSkins) {
 
 test.each(cases)('Feedback %s screen appears properly on %s and %s', async (feedbackType, skin, device) => {
     const page = await openStoryPage({
-        id: `components-screens-feedbackscreen--${feedbackType}`,
+        id: `components-screens-feedbackscreen-${feedbackType}`,
         skin: skin as never,
         device,
     });
@@ -28,7 +32,7 @@ test.each(cases)('Feedback %s screen appears properly on %s and %s', async (feed
 
 test('Feedback screen with large fontSize', async () => {
     const page = await openStoryPage({
-        id: 'components-screens-feedbackscreen--success',
+        id: 'components-screens-feedbackscreen-successfeedbackscreen--success',
         skin: MOVISTAR_SKIN,
         device: 'MOBILE_IOS',
     });