+++ src/navigation-bar.tsx
@@ -226,11 +226,12 @@ const useStyles = createUseStyles((theme) => {
                 isInverse ? theme.colors.navigationBarBackground : theme.colors.background,
             height: DESKTOP_NAVBAR_HEIGHT,
             padding: '16px 0',
-            transition: 'border 300ms',
-            borderBottom: `1px solid ${theme.colors.divider}`,
+            borderBottomStyle: 'solid',
+            borderBottomWidth: ({withBorder}) => (withBorder ? 1 : 0),
             borderColor: ({isInverse}) =>
                 isInverse && !theme.isDarkMode ? 'transparent' : theme.colors.divider,
             [theme.mq.tabletOrSmaller]: {
+                transition: 'border-color 300ms',
                 borderColor: ({isMenuOpen, isInverse}) =>
                     isMenuOpen || (isInverse && !theme.isDarkMode) ? 'transparent' : theme.colors.divider,
                 height: MOBILE_NAVBAR_HEIGHT,
@@ -245,11 +246,7 @@ const useStyles = createUseStyles((theme) => {
             borderBottom: `2px solid transparent`,
             transition: 'border-color 300ms ease-in-out',
 
-            // Only apply hover effect to user agents using fine pointer devices (a mouse, for example)
-            // Also enabled for (pointer: none) for acceptance tests, where (pointer: fine) doesn't match.
-            // WARNING: you may be tempted to use @media (hover: hover) instead, but that doesn't work as expected in some android browsers.
-            // See: https://hover-pointer-media-query.glitch.me/ and https://github.com/mui-org/material-ui/issues/15736
-            '@media (pointer: fine), (pointer: none)': {
+            [theme.mq.supportsHover]: {
                 '&:hover span': {
                     color: ({isInverse}) =>
                         isInverse ? theme.colors.textSecondaryInverse : theme.colors.textSecondary,
@@ -292,11 +289,7 @@ const useStyles = createUseStyles((theme) => {
         },
         iconButton: {
             color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralHigh),
-            // Only apply hover effect to user agents using fine pointer devices (a mouse, for example)
-            // Also enabled for (pointer: none) for acceptance tests, where (pointer: fine) doesn't match.
-            // WARNING: you may be tempted to use @media (hover: hover) instead, but that doesn't work as expected in some android browsers.
-            // See: https://hover-pointer-media-query.glitch.me/ and https://github.com/mui-org/material-ui/issues/15736
-            '@media (pointer: fine), (pointer: none)': {
+            [theme.mq.supportsHover]: {
                 '&:hover': {
                     color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralMedium),
                 },
@@ -334,7 +327,7 @@ export const MainNavigationBar: React.FC<MainNavigationBarProps> = ({
     const [isMenuOpen, setIsMenuOpen] = React.useState(false);
     const [menuTransitionState, setMenuTransitionState] = React.useState<MenuTransitionState>('closed');
     const menuId = useAriaId();
-    const classes = useStyles({isMenuOpen, isInverse, menuTransitionState});
+    const classes = useStyles({isMenuOpen, isInverse, menuTransitionState, withBorder: true});
     const {isTabletOrSmaller} = useScreenSize();
     const setModalState = useSetModalState();
 
@@ -473,6 +466,7 @@ interface NavigationBarCommonProps {
     onBack?: () => void;
     title?: string;
     right?: React.ReactElement;
+    withBorder?: boolean;
     children?: undefined;
 }
 
@@ -495,9 +489,10 @@ export const NavigationBar: React.FC<NavigationBarProps> = ({
     isInverse = false,
     topFixed = true,
     paddingX,
+    withBorder = true,
 }) => {
     const {texts} = useTheme();
-    const classes = useStyles({isInverse, paddingX: paddingX ?? 0});
+    const classes = useStyles({isInverse, paddingX: paddingX ?? 0, withBorder});
     const content = (
         <Inline space="between" alignItems="center">
             <Inline space={24} alignItems="center">
@@ -545,7 +540,7 @@ export const FunnelNavigationBar: React.FC<FunnelNavigationBarProps> = ({
     isInverse = false,
     topFixed = true,
 }) => {
-    const classes = useStyles({isInverse});
+    const classes = useStyles({isInverse, withBorder: true});
     return (
         <ThemeVariant isInverse={isInverse}>
             <header className={classnames(classes.navbar, {[classes.topFixed]: topFixed})}>
@@ -595,11 +590,7 @@ const useNavigationBarActionStyles = createUseStyles((theme) => ({
             color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralHigh),
         },
 
-        // Only apply hover effect to user agents using fine pointer devices (a mouse, for example)
-        // Also enabled for (pointer: none) for acceptance tests, where (pointer: fine) doesn't match.
-        // WARNING: you may be tempted to use @media (hover: hover) instead, but that doesn't work as expected in some android browsers.
-        // See: https://hover-pointer-media-query.glitch.me/ and https://github.com/mui-org/material-ui/issues/15736
-        '@media (pointer: fine), (pointer: none)': {
+        [theme.mq.supportsHover]: {
             '&:hover span': {
                 color: ({isInverse}) =>
                     isInverse ? theme.colors.textSecondaryInverse : theme.colors.textSecondary,